fix all tests and lint
