+++ tests/index.js
@@ -1,4 +1,6 @@
-/*
+/* eslint-disable import/first */
+
+import './utils'
 import 'jss-compose/tests/index.test'
 import 'jss-nested/tests/index.test'
 import 'jss-expand/tests/index.test'
@@ -6,11 +8,7 @@ import 'jss-default-unit/tests/index.test'
 import 'jss-camel-case/tests/index.test'
 import 'jss-extend/tests/index.test'
 import 'jss-isolate/tests/index.test'
-import 'jss-props-sort/tests/index.test'
-import 'jss-vendor-prefixer/tests/index.test'
 import 'react-jss/tests/index.test'
-*/
-
 import './integration/jss'
 import './integration/rules'
 import './integration/sheet'
