+++ tests/utils.js
@@ -1,4 +1,4 @@
-import jss from '../src'
+import jss, {sheets} from '../src'
 
 const {slice} = []
 
@@ -44,9 +44,19 @@ export function computeStyle(className) {
 
 export function reset(jssInstance = jss) {
   jssInstance.plugins.registry = []
+  sheets.reset()
+
+  const styles = document.head.querySelectorAll('[data-jss]')
+  for (let i = 0; i < styles.length; i++) {
+    document.head.removeChild(styles[i])
+  }
 }
 
 export const generateClassName = (str, rule) => (rule.name ? `${rule.name}-id` : 'id')
 
 // Mock the hash function.
 jss.setup({generateClassName})
+
+// Make sure tests are isolated.
+afterEach(reset)
+
